requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tqdm
wandb
SimpleITK
nibabel
ml_collections